refactor: Board 관련 클래스를 board 패키지로 이동
